pyproject.toml
[tool.poetry]
name = "kondo-backend"
version = "0.1.0"
description = ""
authors = ["Ken Erwin <ken@devopslibrary.com>"]

[tool.poetry.dependencies]
cryptography = "^2.6.1"
python = "^3.7"
flask = "^1.0.2"
mypy = "^0.701"
requests = "^2.22.0"
pyjwt = "^1.4.0"
flask-restplus = "^0.12.1"
coverage = "^4.5"
flask-cors = "^3.0"
redis = "^3.2"
gitpython = "^2.1"
testfixtures = "^6.8"
toml = "^0.10.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
black = "^19.3b0"
responses = "^0.10.6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
